Add files via upload
Updated code to allow for analysis of other data sets beyond BCI, and to perform clustering using the louvain algorithm which outperforms the walktrap algorithm. We still get 3 clusters for BCI but the cluster structure is a bit different. We get a tighter relationship with soil but a looser one with traits.